Move bundles.config.js to build-system/compile/
