+++ build-system/tasks/extension-helpers.js
@@ -23,7 +23,7 @@ const {
   extensionAliasBundles,
   extensionBundles,
   verifyExtensionBundles,
-} = require('../../bundles.config');
+} = require('../compile/bundles.config');
 const {compileJs, mkdirSync} = require('./helpers');
 const {endBuildStep} = require('./helpers');
 const {isTravisBuild} = require('../common/travis');
@@ -127,8 +127,8 @@ function declareExtension(
 }
 
 /**
- * Initializes all extensions from bundles.config.js if not already done and
- * populates the given extensions object.
+ * Initializes all extensions from build-system/compile/bundles.config.js if not
+ * already done and populates the given extensions object.
  * @param {?Object} extensionsObject
  * @param {?boolean} includeLatest
  */
